Fix devtab assumption with c->type
Inferno assumed devroot was the first dev in devtab.  They needed
certain chans (ones not associated with a dev yet) to not execute a
dev's close method.  rootclose() was a noop, so they managed.

+++ kern/include/ns.h
@@ -354,7 +354,7 @@ struct chan
 	struct chan*	next;			/* allocation */
 	struct chan*	link;
 	int64_t	offset;			/* in file */
-	uint16_t	type;
+	int	type;
 	uint32_t	dev;
 	uint16_t	mode;			/* read/write */
 	uint16_t	flag;
